Cleaning up CMake and move files into a more structured folder sturcture
