+++ CMakeLists.txt
@@ -30,43 +30,36 @@ MESSAGE("Boost Libraries: ${Boost_LIBRARIES}")
 MESSAGE("Boost Libraries: ${Boost_LOG_LIBRARY}")
 MESSAGE("PROTOBUF INCLUDE DIRS ${PROTOBUF_INCLUDE_DIRS}")
 MESSAGE("PROTOBUF LIBRARIES ${PROTOBUF_LIBRARIES}")
-# IF(Boost_FOUND)
-#   INCLUDE_DIRECTORIES(${Boost_INCLUDE_DIRS})
-#   LINK_DIRECTORIES(${Boost_LIBRARY_DIRS})
-# ENDIF(Boost_FOUND)
 
-# SET(USED_LIBS  ${Boost_LOG_LIBRARY} ${Boost_SYSTEM_LIBRARY} ${Boost_THREAD_LIBRARY} ${Boost_REGEX_LIBRARY})
-
-add_library(libserver
-    "peer.hpp"
-    "header.hpp"
-    "message.hpp"
-    "session.hpp"
-    "message_converter.hpp"
-    "message_queue.cpp"
-    "write_queue.cpp"
-    "session.cpp"
-    "server.hpp"
-    "routing_table.hpp"
-    "message_handler.hpp"
-    "aggregator.hpp"
-    "aggregat.hpp"
-    "cryptowrapper.cpp"
-    "storage.hpp"
-    "proto/messages.pb.cc"
+add_library(libpeerpaste
+    "include/peerpaste/peer.hpp"
+    "include/peerpaste/header.hpp"
+    "include/peerpaste/message.hpp"
+    "src/session.cpp"
+    "include/peerpaste/message_converter.hpp"
+    "src/message_queue.cpp"
+    "src/write_queue.cpp"
+    "include/peerpaste/server.hpp"
+    "include/peerpaste/routing_table.hpp"
+    "include/peerpaste/message_handler.hpp"
+    "include/peerpaste/aggregator.hpp"
+    "include/peerpaste/aggregat.hpp"
+    "src/cryptowrapper.cpp"
+    "include/peerpaste/storage.hpp"
+    "include/peerpaste/proto/messages.pb.cc"
     )
 
-set_target_properties(libserver PROPERTIES LINKER_LANGUAGE CXX)
+set_target_properties(libpeerpaste PROPERTIES LINKER_LANGUAGE CXX)
 
-target_include_directories(libserver
+target_include_directories(libpeerpaste
     PUBLIC
-        # ${BOOST_INCLUDE_DIRS}
-        # ${PROTOBUF_INCLUDE_DIRS}
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+        $<INSTALL_INTERFACE:include>
         ${CRYPTOPP_INCLUDE_DIR}
         "/usr/include/cryptopp"
     )
 
-target_link_libraries(libserver
+target_link_libraries(libpeerpaste
     PUBLIC
         ${CMAKE_THREAD_LIBS_INIT}
         ${Boost_LIBRARIES}
@@ -75,41 +68,7 @@ target_link_libraries(libserver
         ${CRYPTOPP_LIBRARIES}
     )
 
-add_executable(node main.cpp)
-
-target_include_directories(node
-    PUBLIC
-        # ${BOOST_INCLUDE_DIRS}
-)
-
-target_link_libraries(node
-    PUBLIC
-        libserver
-)
-
-add_executable(peer main2.cpp)
-
-target_include_directories(peer
-    PUBLIC
-)
-
-target_link_libraries(peer
-    PUBLIC
-        libserver
-)
-
-add_executable(put main3.cpp)
-
-target_include_directories(put
-    PUBLIC
-)
-
-target_link_libraries(put
-    PUBLIC
-        libserver
-)
-
-add_executable(peerpaste peerpaste.cpp)
+add_executable(peerpaste src/peerpaste.cpp)
 
 target_include_directories(peerpaste
     PUBLIC
@@ -118,44 +77,9 @@ target_include_directories(peerpaste
 target_link_libraries(peerpaste
     PUBLIC
         ${Boost_LIBRARIES}
-        libserver
+        libpeerpaste
 )
-
-# add_executable(msg_test msg_test.cpp)
-
-# target_include_directories(msg_test
-#     PUBLIC
-#         # ${BOOST_INCLUDE_DIRS}
-# )
-
-# target_link_libraries(msg_test
-#     PUBLIC
-#         libserver
-# )
 
 if(BUILD_TESTS)
     add_subdirectory(test)
 endif()
-# find_package(Catch2 REQUIRED)
-
-# add_executable(test test.cpp)
-
-# target_include_directories(test
-#     PUBLIC
-#         ./
-#         # ${BOOST_INCLUDE_DIRS}
-#         # ${PROTOBUF_INCLUDE_DIRS}
-#         # ${CRYPTOPP_INCLUDE_DIR}
-#         # "/usr/include/cryptopp"
-# )
-
-# target_link_libraries(test
-#     PUBLIC
-#         # ${CMAKE_THREAD_LIBS_INIT}
-#         # ${Boost_LIBRARIES}
-#         # # ${USED_LIBS}
-#         # ${Boost_LIBRARY_DIRS}
-#         # ${PROTOBUF_LIBRARIES}
-#         # ${CRYPTOPP_LIBRARIES}
-#         libserver
-# )
