+++ include/peerpaste/cryptowrapper.hpp
