+++ include/peerpaste/message_queue.hpp
