+++ src/cryptowrapper.cpp
@@ -1,4 +1,4 @@
-#include "cryptowrapper.hpp"
+#include "peerpaste/cryptowrapper.hpp"
 
 namespace util {
