+++ test/CMakeLists.txt
@@ -14,6 +14,5 @@ target_include_directories(test
 
 target_link_libraries(test
     PUBLIC
-        # ${PROTOBUF_LIBRARIES}
-        libserver
+        libpeerpaste
 )
